initial exploration
i was able to get access to the original template strings by using
reflection. it's intresting to know there is some sort of barier to
entree for any other projects in the same problem space. also
i am not sure i want to use this feature and publicise it is possible